scst_disk,scst_lib,scst_sysfs,usr/disk_iscsi/scst_disk_session: Add dev_disk_iscsi dev_handler
This device handler is similar to dev_disk.  It is configured to use an
iSCSI disk (open-iscsi), but will initiate a new duplicate iSCSI session
for each client session.

It achieves this by calling up to the new scst-disk-session user-space
utility, which will duplicate the session and then push certain information
down to SCST thru the sysfs interface.